mobile/android/tests/browser/chrome/test_web_channel.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1174458
Migrated from Robocop: https://bugzilla.mozilla.org/show_bug.cgi?id=1184186
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1174458</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/AddTask.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://global/skin"/>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  ChromeUtils.import("resource://gre/modules/Promise.jsm");
  ChromeUtils.import("resource://gre/modules/Services.jsm");
  ChromeUtils.import("resource://gre/modules/XPCOMUtils.jsm");
  ChromeUtils.import("resource://gre/modules/Task.jsm");
  ChromeUtils.defineModuleGetter(this, "WebChannel",
    "resource://gre/modules/WebChannel.jsm");

  const HTTP_PATH = "http://mochi.test:8888";
  const HTTP_ENDPOINT = "/chrome/mobile/android/tests/browser/chrome/web_channel.html";

  const gChromeWin = Services.wm.getMostRecentWindow("navigator:browser");
  let BrowserApp = gChromeWin.BrowserApp;

  // TODO: consider if we want to run the original test in browser-chrome instead

  // Keep this synced with /browser/base/content/test/general/browser_web_channel.js
  // as much as possible.  (We only have this since we can't run browser chrome
  // tests on Android.  Yet?)
  let gTests = [
    {
      desc: "WebChannel generic message",
      run() {
        return new Promise(function(resolve, reject) {
          let tab;
          let channel = new WebChannel("generic", Services.io.newURI(HTTP_PATH));
          channel.listen(function(id, message, target) {
            is(id, "generic");
            is(message.something.nested, "hello");
            channel.stopListening();
            BrowserApp.closeTab(tab);
            resolve();
          });

          tab = BrowserApp.addTab(HTTP_PATH + HTTP_ENDPOINT + "?generic");
        });
      }
    },
    {
      desc: "WebChannel two way communication",
      run() {
        return new Promise(function(resolve, reject) {
          let tab;
          let channel = new WebChannel("twoway", Services.io.newURI(HTTP_PATH));

          channel.listen(function(id, message, sender) {
            is(id, "twoway");
            ok(message.command);

            if (message.command === "one") {
              channel.send({ data: { nested: true } }, sender);
            }

            if (message.command === "two") {
              is(message.detail.data.nested, true);
              channel.stopListening();
              BrowserApp.closeTab(tab);
              resolve();
            }
          });

          tab = BrowserApp.addTab(HTTP_PATH + HTTP_ENDPOINT + "?twoway");
        });
      }
    },
    {
      desc: "WebChannel multichannel",
      run() {
        return new Promise(function(resolve, reject) {
          let tab;
          let channel = new WebChannel("multichannel", Services.io.newURI(HTTP_PATH));

          channel.listen(function(id, message, sender) {
            is(id, "multichannel");
            BrowserApp.closeTab(tab);
            resolve();
          });

          tab = BrowserApp.addTab(HTTP_PATH + HTTP_ENDPOINT + "?multichannel");
        });
      }
    }
  ]; // gTests

  add_task(async function run_all() {
    for (let test of gTests) {
      info("Running: " + test.desc);
      await test.run();
    }
  });

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1174458">Mozilla Bug 1174458</a>
<br>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1184186">Migrated from Robocop testWebChannel</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
